Add files via upload
Tables
https://dev.mysql.com/doc/sakila/en/sakila-structure-tables.html
Exercises


Display the first and last name of each actor in a single column in upper case letters in alphabetic order. Name the column Actor Name.

 Find all actors whose last name contain the letters GEN:

Using IN, display the country_id and country columns of the following countries: Afghanistan, Bangladesh, and China:
 List the last names of actors, as well as how many actors have that last name.

List last names of actors and the number of actors who have that last name, but only for names that are shared by at least two actors

The actor HARPO WILLIAMS was accidentally entered in the actor table as GROUCHO WILLIAMS. Write a query to fix the record.
 Use JOIN to display the first and last names, as well as the address, of each staff member. Use the tables staff and address:
List each film and the number of actors who are listed for that film. Use tables film_actor and film. Use inner join.
 How many copies of the film Hunchback Impossible exist in the inventory system?
Using the tables payment and customer and the JOIN command, list the total paid by each customer. List the customers alphabetically by last name
The music of Queen and Kris Kristofferson have seen an unlikely resurgence. As an unintended consequence, films starting with the letters K and Q have also soared in popularity. Use subqueries to display the titles of movies starting with the letters K and Q whose language is English.
Use subqueries to display all actors who appear in the film Alone Trip.
You want to run an email marketing campaign in Canada, for which you will need the names and email addresses of all Canadian customers. Use joins to retrieve this information.
Sales have been lagging among young families, and you wish to target all family movies for a promotion. Identify all movies categorized as famiy films.
Create a Stored procedure to get the count of films in the input category (IN category_name, OUT count)
Display the most frequently rented movies in descending order.
Write a query to display for each store its store ID, city, and country.
 List the genres and its gross revenue. 
Create a View for the above query(18)
Select top 5  genres in gross revenue view.